requirements.txt
python-telegram-bot==22.5
peewee==3.18.3
pytest==8.4.2
python-dotenv==1.2.1